New: added release-it
Added release-it with conventional-changelog and eslint preset
Removed auto-changelog
Updated README
Moved previous version post bump scripts to release-it plugin
