+++ .release-it.json
@@ -1,0 +1,29 @@
+{
+  "default": "https://github.com/release-it/release-it/blob/master/conf/release-it.json",
+  "hooks": {
+    "before:init": "npm run lint"
+  },
+  "git": {
+    "requireCleanWorkingDir": false,
+    "requireCommits": false,
+    "tagName": "v${version}",
+    "commit": true,
+    "push": true
+  },
+  "npm": {
+    "publish": false
+  },
+  "github": {
+    "release": true,
+    "releaseName": "Release ${version}",
+    "draft": true,
+    "tokenRef": "GITHUB_ACCESS_TOKEN"
+  },
+  "plugins": {
+    "./scripts/bump-version.js": {},
+    "@release-it/conventional-changelog": {
+      "preset": "eslint",
+      "infile": "CHANGELOG.md"
+    }
+  }
+}
